+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.1.1
+##  03/05/2022
+
+1. [](#new)
+    * Store Nick in Session var for Convenience (remain 'logged in' during Session)
+    * add Comments Function
+
 # v0.1.0
 ##  26/04/2022